# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API สำหรับระบบ Life Long Education ที่พัฒนาด้วย Go Fiber",
        "title": "Life Long Education API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/list-queue": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ListQueue"
                ],
                "summary": "ดึงคำร้องทั้งหมด",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.ListQueue"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ListQueue"
                ],
                "summary": "แก้ไขคำร้อง",
                "parameters": [
                    {
                        "description": "ListQueue Update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.ListQueue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ListQueue"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ListQueue"
                ],
                "summary": "สร้างคำร้องใหม่",
                "parameters": [
                    {
                        "description": "ListQueue Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.ListQueue"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ListQueue"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/list-queue/staff-status/{staff_status_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ListQueue"
                ],
                "summary": "ดึงคำร้องตาม Staff Status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Staff Status ID",
                        "name": "staff_status_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.ListQueue"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/list-queue/{id}/staff-status/{staff_status_id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ListQueue"
                ],
                "summary": "อัปเดตสถานะของเจ้าหน้าที่",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ListQueue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Staff Status ID",
                        "name": "staff_status_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ListQueue"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.Faculty": {
            "type": "string",
            "enum": [
                "คณะศึกษาศาสตร์",
                "คณะมนุษยศาสตร์",
                "คณะวิจิตรศิลป์",
                "คณะสังคมศาสตร์",
                "คณะวิทยาศาสตร์",
                "คณะวิศวกรรมศาสตร์",
                "คณะแพทยศาสตร์",
                "คณะเกษตรศาสตร์",
                "คณะทันตแพทยศาสตร์",
                "คณะเภสัชศาสตร์",
                "คณะเทคนิคการแพทย์",
                "คณะพยาบาลศาสตร์",
                "คณะอุตสาหกรรมเกษตร",
                "คณะสัตวแพทยศาสตร์",
                "คณะสถาปัตยกรรมศาสตร์",
                "คณะบริหารธุรกิจ",
                "คณะเศรษฐศาสตร์",
                "คณะนิติศาสตร์",
                "คณะสื่อสารมวลชน",
                "คณะรัฐศาสตร์",
                "วิทยาลัยศิลปะ สื่อ และเทคโนโลยี",
                "คณะสาธารณสุขศาสตร์",
                "วิทยาลัยการศึกษาและการจัดการทะเล",
                "วิทยาลัยนานาชาติเทคโนโลยีดิจิทัล",
                "สถาบันนโยบายสาธารณะ",
                "สถาบันวิศวกรรมชีวการแพทย์",
                "สำนักบริการวิชาการ",
                "สำนักงานพัฒนาคุณภาพการศึกษา",
                "สำนักงานมหาวิทยาลัย",
                "ศูนย์วิจัยข้าวล้านนา",
                "สถาบันภาษา",
                "วิทยาลัยการศึกษาตลอดชีวิต",
                "ศูนย์บริหารจัดการความปลอดภัยฯ SHE",
                "ศูนย์นวัตกรรมการสอนและการเรียนรู้ TLIC"
            ],
            "x-enum-varnames": [
                "FacultyOfEducation",
                "FacultyOfHumanities",
                "FacultyOfFineArts",
                "FacultyOfSocialSciences",
                "FacultyOfScience",
                "FacultyOfEngineering",
                "FacultyOfMedicine",
                "FacultyOfAgriculture",
                "FacultyOfDentistry",
                "FacultyOfPharmacy",
                "FacultyOfMedTechnology",
                "FacultyOfNursing",
                "FacultyOfAgroIndustry",
                "FacultyOfVeterinaryMedicine",
                "FacultyOfArchitecture",
                "BachelorOfBussinessAdministration",
                "FacultyOfEconomics",
                "FacultyOfLaw",
                "FacultyOfMassCommunication",
                "FacultyOfPoliticalScience",
                "CollegeOfArtMediaAndTechnology",
                "FacultyOfPublicHealth",
                "CollegeOfEducationAndSeaManagement",
                "InternationalCollegeOfDigitalTechnology",
                "InstituteOfPublic",
                "BiomedicalEngineeringInstitute",
                "AcademicServiceCenter",
                "OfficeOfEducationQualityDevelopment",
                "OfficeOfTheUniversity",
                "LannaRiceResearchCenter",
                "LanguageInstitute",
                "SchoolOfLifeLongEducation",
                "CenterOfSafetyOccupationalHealthAndEnvironment",
                "TeachingAndLearningInnovationCenter"
            ]
        },
        "entities.ListQueue": {
            "type": "object",
            "properties": {
                "appointment_date_aw": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_left": {
                    "description": "ใช้ pointer เพื่อรองรับ null",
                    "type": "integer"
                },
                "faculty": {
                    "$ref": "#/definitions/entities.Faculty"
                },
                "id": {
                    "type": "integer"
                },
                "info_submit": {
                    "type": "string"
                },
                "info_submit_14days": {
                    "type": "string"
                },
                "on_web": {
                    "type": "string"
                },
                "order_mappings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.OrderMapping"
                    }
                },
                "priority": {
                    "type": "integer"
                },
                "staff_id": {
                    "type": "integer"
                },
                "staff_status": {
                    "$ref": "#/definitions/entities.StaffStatus"
                },
                "staff_status_id": {
                    "type": "integer"
                },
                "time_register": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "users_status": {
                    "$ref": "#/definitions/entities.UserStatus"
                },
                "users_status_id": {
                    "type": "integer"
                },
                "wordfile_submit": {
                    "type": "string"
                }
            }
        },
...
        "entities.OrderMapping": {
            "type": "object",
            "properties": {
                "checked": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "list_queue_id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.StaffStatus": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "entities.UserStatus": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token สำหรับการยืนยันตัวตน ให้ใส่ token ในรูปแบบ: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}